[MLIR][SPIRV] Add a ()-pair around literal struct fields.
Summary:

In order to support recursive structs in the SPIRV dialect, we are going
to adapt the proposal detailed in [1]. As a first step towards that
goal, literal (non-identified) structs need to be modified such their
fields are surrounded by a ()-pair. This commit introduces such change
to the syntax.

[1] https://llvm.discourse.group/t/rfc-first-party-support-for-llvm-types/1295