[MLIR][SPIRV] Support identified and recursive structs.
This PR adds support for identified and recursive structs. This includes: parsing,
printing, serializing, and deserializing such structs.

The following C struct:

```C
struct A {
  A* next;
};
```

which is translated to the following MLIR code as:

```mlir
!spv.struct<A, (!spv.ptr<!spv.struct<A>, Generic>)>
```

would be represented in the SPIR-V module as:

```spirv
OpName %A "A"
OpTypeForwardPointer %APtr Generic
%A = OpTypeStruct %APtr
%APtr = OpTypePointer Generic %A
```

In particular the following changes are included:
- SPIR-V structs can now be either identified or literal (i.e. non-identified).
- All structs now have their members surrounded by a ()-pair.
- For recursive references, (1) an OpTypeForwardPointer instruction is emitted before
the OpTypeStruct instruction defining the recursive struct, and (2) an OpTypePointer
instruction is emitted after the OpTypeStruct instruction which actually defines the
recursive pointer to struct type.

Differential Revision: https://reviews.llvm.org/D87206