[MLIR][SPIRV] Add SPV_StructuredBranchOp to model structured CF.
Summary:

This patches is the first step towards re-writing SPIR-V dialect's
(de-)serialization logic. The final goal is to have lower-level
representation of SPIR-V programs within MLIR. The main objective
here is to have an intermediate step within MLIR that more faithfully
resembles the control flow structure required by the SPIR-V binary format.
At the moment, there is a big jump between MLIR region-based high-level CF
ops and SPIR-V's low-level CF structure as defined by OpSelectionMerge and
OpLoopMerge during (de-)serialization. The downside for this is that it
might be difficult to verify or troubleshoot (de-)serialization logic when
going between the 2 representations. Therefore, the need for the
intermediate step that lives in the MLIR world but still close in
representation to the low-level binary world.

To that end, this patch introduces the following changes:

- adds a rewrite pattern that converts spv.mlir.loop into their
  spv.mlir.structured_br counterparts.
- a new SPV_StructuredBranchOp that represents a combination of a SPIR-V
  merge op + a branch op.